Working with policies

+++ app/Http/Controllers/ThreadController.php
@@ -118,6 +118,12 @@ class ThreadController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+
+//        if ($thread->user_id != auth()->id()) {
+//            abort(403, 'You do not have permission to do this');
+//        }
+
+        $this->authorize('update', $thread);
         $thread->replies()->delete();
         $thread->delete();
